moved app test and customer cart test to Testing folder
