+++ src/Testing/App.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import App from "./App";
+import App from "../App";
 import { HashRouter } from "react-router-dom";
 
 test("Contains two routes", () => {
